pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = ""
authors = [
    {name = "Jerenemy",email = "jzay@wesleyan.edu"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "torchaudio (>=2.9.1,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "torchcodec (>=0.9.1,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
